+++ flash-core/src/main/java/cn/enilu/flash/service/message/sms/tencent/TencentSmsSender.java
@@ -3,6 +3,7 @@ package cn.enilu.flash.service.message.sms.tencent;
 import cn.enilu.flash.bean.enumeration.ConfigKeyEnum;
 import cn.enilu.flash.cache.ConfigCache;
 import cn.enilu.flash.service.message.sms.SmsSender;
+import cn.enilu.flash.utils.Lists;
 import cn.enilu.flash.utils.StringUtil;
 import com.github.qcloudsms.SmsSingleSender;
 import com.github.qcloudsms.SmsSingleSenderResult;
@@ -11,6 +12,9 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
+import java.util.LinkedHashMap;
+
 @Service
 public class TencentSmsSender implements SmsSender {
     private Logger logger = LoggerFactory.getLogger(TencentSmsSender.class);
@@ -18,7 +22,7 @@ public class TencentSmsSender implements SmsSender {
     private ConfigCache configCache;
 
     @Override
-    public boolean sendSms(String tplCode, String receiver, String[] args, String content) {
+    public boolean sendSms(String tplCode, String receiver, LinkedHashMap params, String content) {
         Integer appid = Integer.valueOf((String) configCache.get(ConfigKeyEnum.API_TENCENT_SMS_APPID.getValue()));
         String appkey = (String) configCache.get(ConfigKeyEnum.API_TENCENT_SMS_APPKEY.getValue());
         String smsSign = (String) configCache.get(ConfigKeyEnum.API_TENCENT_SMS_SIGN.getValue());
@@ -26,10 +30,10 @@ public class TencentSmsSender implements SmsSender {
         SmsSingleSenderResult result = null;
         try {
             if (StringUtil.isNotEmpty(tplCode)) {
+                ArrayList list = (ArrayList) Lists.newArrayList(params.values());
                 //根据指定模板id发送短信
-                // 签名参数未提供或者为空时，会使用默认签名发送短信
                 result = ssender.sendWithParam("86", receiver,
-                        Integer.valueOf(tplCode), args, smsSign, "", "");
+                        Integer.valueOf(tplCode), list, smsSign, "", "");
             } else {
                 //发送固定内容短信
                 result = ssender.send(0, "86", receiver,